Move FontSizeControl to the `ui` subfolder
