+++ packages/components/src/ui/font-size-control/index.js
@@ -1,0 +1,2 @@
+export { default as FontSizeControl } from './font-size-control';
+export { default as useFontSizeControl } from './use-font-size-control';
