+++ packages/components/src/ui/truncate/styles.js
