+++ packages/components/src/ui/truncate/utils.js
