+++ packages/components/tsconfig.json
@@ -29,6 +29,7 @@
 		"src/**/*.native.js",
 		"src/**/react-native-*",
 		"src/**/stories",
-		"src/**/test"
+		"src/**/test",
+		"src/ui/font-size-control"
 	]
 }
